Administration: Remove hiding certain UI elements
We had the ability to hide certain parts of the UI if they weren't used.
However, this was implemented inconsistently, both in terms of what features
could be hidden and hiding the features properly. This removes the feature,
since it is straightforward to ignore UI elements unrelated to your current
task.

This also removes SettingsForm, since the reverse_checkboxes it provided were
only used in hiding UI elements.

refs #10447