WIP WIP WIP
Re-use ReplacementDict model which was unused in order to store local config
or replacements for tasks in StandardTaskConfig. Previously, AtoM config was
written directly in StandardTaskConfig. ATK/AS config was written in
MicroServiceChoiceReplacementDic. These are the only cases where we had to make
these changes.

The goal is to facilitate the task of pulling out workflow data from the
database. ReplacementDict model becomes the new way to provide replacements.

TODO:
- Review forms: initial defaults (when missing in RD), validation...
- Migrate old Config models in administration.models (migrations)
- Migrate AtoM MSCL task to RD instead of run-once
- Update --debug in upload-quibt to take string
- Update ReplaceDic link task manager ot use new Replace table
- Tests!